+++ Changelog.md
@@ -75,3 +75,10 @@
 - Made the dropout in each layer controllable
 - Added new options to 'train_model' script
 - Extended the full backward diffusion at validation with an additional run for a sample from the train dataset
+
+### 1.3.0
+
+- Added script for splitting processed samples into chunks
+- Made spectrograms normalizing optional
+- Modified Encoder's architecture
+- Fixed bug with fixed Decoder's parameters